Authentication

+++ api/courses.js
@@ -1,10 +1,10 @@
-const { Router, application } = require('express')
-const sequelize = require
+const { Router } = require('express')
 const { Course } = require('../models/course')
 const { Enrollment } = require('../models/enrollment')
-const { User } = require('../models/user')
+const { User, requireInstructorMatchesBody, requireInstructorMatchesParams } = require('../models/user')
 const { Assignment } = require('../models/assignment')
 const { validateAgainstSchema } = require('../lib/validation')
+const { requireAuthentication } = require('../lib/auth')
 
 const router = Router()
 
@@ -18,7 +18,7 @@ courseSchema = {
 
 enrollmentSchema = {
     add: { required: true },
-    delete: { required: true }
+    remove: { required: true }
 }
 
 /*
@@ -77,40 +77,48 @@ router.get('/', async function (req, res) {
 })
 
 router.get('/:id', async function (req, res, next) {
-    const courseId = parseInt(req.params.id)
-    const course = await Course.findByPk(courseId)
-    if (course) {
+    try {
+        const courseId = parseInt(req.params.id)
+        const course = await Course.findByPk(courseId)
         if (course) {
-        res.status(200).send(course)
-    } else {
-        next()
+            if (course) {
+                res.status(200).send(course)
+            } else {
+                next()
+            }
+        } else {
+            res.status(404).send({
+                error: "no course with that id was found"
+            })
+        }
+    } catch (e) {
+        next(e)
     }
-    } else {
-        res.status(400).send({
-            error: "no course with that id was found"
-        })
-    }
-    
 })
 
-router.get('/:id/students', async function (req, res) {
-    const courseId = parseInt(req.params.id)
-    const enrollment = await Enrollment.findAll({ where: { courseId: courseId } })
-    if (enrollment.length > 0) {
-        const users = await User.findAll()
-        const id = []
-        for (let i = 0; i < enrollment.length; i++) {
-            id[id.length] = enrollment[i].studentId
+router.get('/:id/students', requireAuthentication, requireInstructorMatchesParams, async function (req, res, next) {
+    try {
+        const courseId = parseInt(req.params.id)
+        const enrollment = await Enrollment.findAll({ where: { courseId: courseId } })
+        if (enrollment.length > 0) {
+            const users = await User.findAll()
+            const id = []
+            for (let i = 0; i < enrollment.length; i++) {
+                id[id.length] = enrollment[i].studentId
+            }
+            const students = isStudent(id, users)
+            res.status(200).send({
+                students: students
+            })
+        } else {
+            res.status(404).send({
+                error: "no course with that id found"
+            })
         }
-        const students = isStudent(id, users)
-        res.status(200).send({
-            students: students
-        })
-    } else {
-        res.status(404).send({
-            error: "no course with that id found"
-        })
+    } catch (e) {
+        next(e)
     }
+
 })
 
 function isStudent(num, arr) {
@@ -127,100 +135,120 @@ function isStudent(num, arr) {
     return students
 }
 
-//validate user
-router.get('/:id/roster', async function (req, res) {
-    const courseId = parseInt(req.params.id)
-    const enrollment = await Enrollment.findAll({ where: { courseId: courseId } })
-    if (enrollment.length > 0) {
-        const users = await User.findAll()
-        const id = []
-        for (let i = 0; i < enrollment.length; i++) {
-            id[id.length] = enrollment[i].studentId
-        }
-        const students = isStudent(id, users)
-        let csv = "name, studentId, email \n"
-        for (let i = 0; i < students.length; i++) {
-            csv += `${students[i].name}, ${students[i].id}, ${students[i].email}\n`
-        }
-        res.setHeader('Content-Type', 'text/csv')
-        res.setHeader('Content-Disposition', 'attachment; filename-roster.csv')
-        res.status(200).send(csv)
-    } else {
-        res.status(404).send({
-            error: "no course with that id found"
-        })
-    }
-
-})
-
-router.get('/:id/assignments', async function (req, res) {
-    const courseId = parseInt(req.params.id)
-    const assignment = await Assignment.findAll({ where: { courseId: courseId } })
-    if (assignment.length > 0) {
-        res.status(200).send({
-            assignments: assignment
-        })
-    } else {
-        res.status(400).send({
-            error: "no course found with that id"
-        })
-    }
-})
-
-router.post('/', async function (req, res, next) {
-    const request = await req.body
-    if (validateAgainstSchema(request, courseSchema)) {
-        const course = await Course.create(request)
-        console.log("course: ", course.toJSON())
-        res.status(201).send({
-            id: course.id,
-            links: {
-                course: `/course/${course.id}`
-            }
-        })
-    } else {
-        res.status(400).send({
-            error: "body is not a valid course object"
-        })
-    }
-})
-
-router.post('/:id/students', async function (req, res, next) {
-    const request = await req.body
-    if (validateAgainstSchema(request, enrollmentSchema)) {
+router.get('/:id/roster', requireAuthentication, requireInstructorMatchesParams, async function (req, res, next) {
+    try {
         const courseId = parseInt(req.params.id)
-        const add = request.add
-        const remove = request.delete
-        const course = await Course.findByPk(courseId)
-        if (course) {
-            for (let i = 0; i < add.length; i++) {
-                await Enrollment.create({
-                    studentId: add[i],
-                    courseId: courseId
-                })
+        const enrollment = await Enrollment.findAll({ where: { courseId: courseId } })
+        if (enrollment.length > 0) {
+            const users = await User.findAll()
+            const id = []
+            for (let i = 0; i < enrollment.length; i++) {
+                id[id.length] = enrollment[i].studentId
             }
-            for (let i = 0; i < remove.length; i++) {
-                await Enrollment.destroy({
-                    where: {
-                        studentId: remove[i],
-                        courseId: courseId
-                    }
-                })
+            const students = isStudent(id, users)
+            let csv = "name, studentId, email \n"
+            for (let i = 0; i < students.length; i++) {
+                csv += `${students[i].name}, ${students[i].id}, ${students[i].email}\n`
             }
-            res.status(200).send()
+            res.setHeader('Content-Type', 'text/csv')
+            res.setHeader('Content-Disposition', 'attachment; filename-roster.csv')
+            res.status(200).send(csv)
         } else {
             res.status(404).send({
-                error: "course not found"
+                error: "no course with that id found"
             })
         }
-    } else {
-        res.status(400).send({
-            error: "body is not a valid object"
-        })
+    } catch (e) {
+        next(e)
     }
 })
 
-router.patch('/:id', async function (req, res, next) {
+router.get('/:id/assignments', async function (req, res, next) {
+    try {
+        const courseId = parseInt(req.params.id)
+        const assignment = await Assignment.findAll({ where: { courseId: courseId } })
+        if (assignment.length > 0) {
+            res.status(200).send({
+                assignments: assignment
+            })
+        } else {
+            res.status(404).send({
+                error: "no course found with that id"
+            })
+        }
+    } catch (e) {
+        next(e)
+    }
+})
+
+router.post('/', requireAuthentication, async function (req, res, next) {
+    try {
+        const request = await req.body
+        if (req.role == 'admin') {
+            if (validateAgainstSchema(request, courseSchema)) {
+                const course = await Course.create(request)
+                console.log("course: ", course.toJSON())
+                res.status(201).send({
+                    id: course.id,
+                    links: {
+                        course: `/course/${course.id}`
+                    }
+                })
+            } else {
+                res.status(400).send({
+                    error: "body is not a valid course object"
+                })
+            }
+        } else {
+            res.status(403).send({
+                error: "must be admin"
+            })
+        }
+    } catch (e) {
+        next(e)
+    }
+})
+
+router.post('/:id/students', requireAuthentication, requireInstructorMatchesParams, async function (req, res, next) {
+    try {
+        const courseId = parseInt(req.params.id)
+        const request = await req.body
+        const add = request.add
+        const remove = request.remove
+        const course = await Course.findByPk(courseId)
+        if (validateAgainstSchema(request, enrollmentSchema)) {
+            if (course) {
+                for (let i = 0; i < add.length; i++) {
+                    await Enrollment.create({
+                        studentId: add[i],
+                        courseId: courseId
+                    })
+                }
+                for (let i = 0; i < remove.length; i++) {
+                    await Enrollment.destroy({
+                        where: {
+                            studentId: remove[i],
+                            courseId: courseId
+                        }
+                    })
+                }
+                res.status(200).send()
+            } else {
+                res.status(404).send({
+                    error: "course not found"
+                })
+            }
+        } else {
+            res.status(400).send({
+                error: "body is not a valid object"
+            })
+        }
+    } catch (e) {
+        next(e)
+    }
+})
+
+router.patch('/:id', requireAuthentication, requireInstructorMatchesParams, async function (req, res, next) {
     const request = await req.body
     const courseId = parseInt(req.params.id)
     const course = await Course.findByPk(courseId)
@@ -242,17 +270,27 @@ router.patch('/:id', async function (req, res, next) {
     }
 })
 
-router.delete('/:id', async function (req, res, next) {
-    const courseId = parseInt(req.params.id)
-    const course = await Course.findByPk(courseId)
-    if (course) {
-        const result = await Course.destroy({ where: { id: course.id } })
-        if (result > 0) {
-            res.status(204).send()
+router.delete('/:id', requireAuthentication, async function (req, res, next) {
+    try {
+        const courseId = parseInt(req.params.id)
+        const course = await Course.findByPk(courseId)
+        if (req.role == 'admin') {
+            if (course) {
+                const result = await Course.destroy({ where: { id: course.id } })
+                if (result > 0) {
+                    res.status(204).send()
+                }
+                console.log(result)
+            } else {
+                next()
+            }
+        } else {
+            res.status(403).send({
+                error: "must be admin"
+            })
         }
-        console.log(result)
-    } else {
-        next()
+    } catch (e) {
+        next(e)
     }
 })
